Add support for providers with no WordPress registered settings. This fixes a bug where "Save Settings" was shown, even though there were no settings.
Add basic CSV provider. This adds support for very basic CSVs where line 1 contains field keys, and all other lines contain objects. column 1 of the CSV must be a unique identifier (which is also used as the title of the object in WordPress admin). Field keys are not processed or slugified (but maybe they should be?) so you need to use the full case sensitive version. Version 3 of CultureObject will contain a much more complicated version of this provider and will allow for a "schema" version of the CSV to be saved so we can support remapping of fields, as well as custom fields being used for the title, and the unique object identifier.